fix error double quotes in sqlite3
This commit addresses an issue arising from the addition of double quotes in a previous commit. These double quotes were added to refer to column names containing spaces or hyphens. However, it has been observed that if the referred column does not exist, SQLite3 interprets the quoted "name" as a literal string, not a column reference. This causes SQLite3 to return a column named "name" repeated by the length of the table. This commit contains fixes for this issue.

* fix: replace double quotes with backtick
* update all the tests to match the new requirements